AVBYTB-5 implemented up-to-date car data receiving
