+++ src/main/java/com/github/av/bytelegrambot/Application.java
@@ -2,7 +2,9 @@ package com.github.av.bytelegrambot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableScheduling
 @SpringBootApplication
 public class Application {
