+++ src/main/java/com/github/av/bytelegrambot/service/DatabaseUpdateService.java
@@ -1,0 +1,8 @@
+package com.github.av.bytelegrambot.service;
+
+public interface DatabaseUpdateService {
+
+    void initDB();
+
+    void updateDB();
+}
