+++ src/main/resources/application.properties
@@ -1,5 +1,6 @@
 bot.username=avby_tgbot
 bot.token=5795980201:AAGEfj3tUOv3bLWQzZDaniuVbzNAMeb2ORE
+bot.databaseUpdateFixedRate=86400000
 
 spring.messages.basename=lang/res
